+++ app/frontend/src/javascript/components/authentication-provider/saml-form.tsx
@@ -1,12 +1,13 @@
 import { FormInput } from '../form/form-input';
-import { UseFormRegister, FormState } from 'react-hook-form';
+import { FormSwitch } from '../form/form-switch';
+import { UseFormRegister, FormState, Control } from 'react-hook-form';
 import { FieldValues } from 'react-hook-form/dist/types/fields';
 import { useTranslation } from 'react-i18next';
 import { FabOutputCopy } from '../base/fab-output-copy';
-import ValidationLib from '../../lib/validation';
 
-interface SamlFormProps<TFieldValues> {
+interface SamlFormProps<TFieldValues, TContext extends object> {
   register: UseFormRegister<TFieldValues>,
+  control: Control<TFieldValues, TContext>,
   formState: FormState<TFieldValues>,
   strategyName?: string,
 }
@@ -14,7 +15,7 @@ interface SamlFormProps<TFieldValues> {
 /**
  * Partial form to fill the OAuth2 settings for a new/existing authentication provider.
  */
-export const SamlForm = <TFieldValues extends FieldValues>({ register, strategyName, formState }: SamlFormProps<TFieldValues>) => {
+export const SamlForm = <TFieldValues extends FieldValues, TContext extends object>({ register, strategyName, formState, control }: SamlFormProps<TFieldValues, TContext>) => {
   const { t } = useTranslation('admin');
 
   /**
@@ -39,7 +40,7 @@ export const SamlForm = <TFieldValues extends FieldValues>({ register, strategyN
                  placeholder="https://sso.example.net..."
                  label={t('app.admin.authentication.saml_form.idp_sso_service_url')}
                  tooltip={t('app.admin.authentication.saml_form.idp_sso_service_url_help')}
-                 rules={{ required: true, pattern: ValidationLib.urlRegex }}
+                 rules={{ required: true }}
                  formState={formState} />
       <FormInput id="providable_attributes.idp_cert_fingerprint"
                  register={register}
@@ -56,14 +57,29 @@ export const SamlForm = <TFieldValues extends FieldValues>({ register, strategyN
                  placeholder="https://exemple.net/user..."
                  label={t('app.admin.authentication.saml_form.profile_edition_url')}
                  tooltip={t('app.admin.authentication.saml_form.profile_edition_url_help')}
-                 rules={{ required: true, pattern: ValidationLib.urlRegex }}
+                 rules={{ required: true }}
                  formState={formState} />
       <FormInput id="providable_attributes.idp_slo_service_url"
                  register={register}
                  placeholder="https://sso.exemple.net..."
                  label={t('app.admin.authentication.saml_form.idp_slo_service_url')}
                  tooltip={t('app.admin.authentication.saml_form.idp_slo_service_url_help')}
-                 rules={{ pattern: ValidationLib.urlRegex }}
+                 formState={formState} />
+      <FormSwitch id="providable_attributes.authn_requests_signed" control={control}
+                  formState={formState}
+                  label={t('app.admin.authentication.saml_form.authn_requests_signed')} />
+      <FormSwitch id="providable_attributes.want_assertions_signed" control={control}
+                  formState={formState}
+                  label={t('app.admin.authentication.saml_form.want_assertions_signed')} />
+      <FormInput id="providable_attributes.sp_certificate"
+                 register={register}
+                 placeholder="-----BEGIN CERTIFICATE-----...-----END CERTIFICATE-----"
+                 label={t('app.admin.authentication.saml_form.sp_certificate')}
+                 formState={formState} />
+      <FormInput id="providable_attributes.sp_private_key"
+                 register={register}
+                 placeholder="-----BEGIN RSA PRIVATE KEY-----...-----END RSA PRIVATE KEY-----"
+                 label={t('app.admin.authentication.saml_form.sp_private_key')}
                  formState={formState} />
     </div>
   );