+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - improvement: add loader for create/delete availability slot
 - improvement: allow admin configure memeber's profile gender/birthday as required
+- improvement: add sp certificate for saml provider
 - Fix a bug: unable to update a space with a deleted machine
 - Fix a bug: unable to get invoice payment details if the account code is same for card/transfer payment method
 - updates translations